requirements.txt
pymongo==4.2.0
pika==1.3.0
requests==2.28.1
beautifulsoup4==4.11.1
git+https://github.com/codelucas/newspaper.git
python-dateutil==2.8.2
structlog==22.1.0
pytz==2022.2.1
